Add the launching of the UI tests to mainBuildSteps.groovy.
Note that there is a Selenium/Chrome issue which requires the
test suite to not be launched with sudo.  However, Hermes has
to be launched with sudo to delete any existing DB files. So
instead, we will delete the DB files before launching the UI
tests, then launch the UI tests without sudo.