VB-286: upgrade ubuntu to fix valgrind/memleak test (#90)
The new state transfer code uses a randomize function that attempts to
execute an RDRAND opcode. There is a bug in valgrind 3.11 that
advertises this opcode as available, even though they don't support
it. Unfortunately, that's the newest valgrind available on Ubuntu
16.04. The bug was fixed in valgrind 3.12. Upgrading to Ubuntu 18.04
brings in valgrind 3.13.

This patch updates the prereqs build image and also the run image, since we need to link to the correct versions of things like protobuf. It also updates the dependency on concord-bft to pick up changes needed to address warnings issued by gcc 7.