Containerize all Maven builds and revise README
The patch changes the way ethrpc and helen build in CICD, and
updates the README to reflect the revised version of build steps.

This patch also includes an update in README to  includes an
alternative build step for developers who do not have Java-based
toolchain installed. (Containerized build)

As a result, both ethrpc and helen now utilize a fully containerized
build. To help with developers actively developing on these projects,
the patch does include packaging-only Dockerfiles for both projects
that only containerize a locally pre-built jar into respective docker
images.