Add helen config to serve new ethrpc certs and correct ports
This adds the new certs to their proper place in the
/api/concord/members response.

Note: you'll need to clean any pre-existing helen database entries, or
the new default rpc cert config will be ignored.

The change to ethrpc ports in docker-compose is because the ethrpc jar
always starts on 8545, so the container port must always be 8545. Only
the external port is routed to not conflict with other ethrpc
instances.