Readme cleanup (#126)
* Readme cleanup

We haven't really scrubbed our readmes since the move to the
monorepo. This is the first step of that task. This change proposes
the following:

We have three levels of documentation:

1. https://confluence.eng.vmware.com/display/BLOC/Docker explains how
to start up all of our services using *prebuilt* docker images. (This
is TODO at the moment, and will happen shortly after this PR opens.)
It asks people to refer to the top-level README in this repo for
instructions on how to actually build their own docker images.

2. The top-level README in this repo describes how to build everything
in this repository as docker images. Ideally, there is one section per
sub-project, and that section is not much more than a single "docker
build ...".

3. Each subproject has a README that describes the steps needed to
build the project outside of docker.

In this structure, the Confluence page gets those with casual interest
up and running, the top-level readme gets developers a basic running
system, and the individual sub-readmes help anyone actually working on
that specific subprojects code.

Major changes included here:

1. Adding some notes to each readme suggesting the structure described
above.

2. Removing UI build instructions from the Helen readme.

3. Removing old non-monorepo links from all readmes.

4. Restructuring the non-docker sample configs for concord so that the
readme instructions work without having to move genesis.json around
yourself.