VB-221: Log Intelligence (LINT) Passthrough
I've implemented a LINT Proxy Controller.

Any call made to /api/lint/.... will be proxied to LINT, and the
response will be returned.

For example, to make the call
https://api.mgmt.cloud.vmware.com/ops/query/log-query-tasks/

Call helen with the endpoint
https://helen/api/lint/ops/query/log-query-tasks

The proxy does the following:
For any operation, add the lint service user Bearer token to the
Authorization header.  Thus, all calls to LINT are done as the service
user.

If there is a parameter in the body called "logQuery", add a filter to
limit to the caller's consortiumId, and the specified function.

The default function filter is "ethLogger", but may be changed in a
queryParm: https://helen/api/lint/ops/query/log-query-tasks?function=helenLogger

There is a basic pattern matcher for the sql query, so it's probably
pretty easy to break.

Unit tests were also added.