move docker directory out of concord (#132)
The example docker-compose configuration has been in concord/docker/
only because it was in the concord repo before the monorepo
merge. Since it contains information about the rest of the Athena
services, it is much more appropriate to keep it at the top-level of
this repository, at docker/.

Also remove helen/docker-compose.yml, which is around for a similar
reason, but has become defunct.