Modify default value generation
This patch modifies the default value generation to be more advanced
than simple literals (using CodeBlock construct instead). This allows
the generated code to be better type-aware.

This patch fixes a bug where generated default value for nested enum was
not handled properly (the generated default value statement for inner
enum was not accounting for nested class/type hierarchy).

This patch also adds a .proto message for 'oneof', which exercises
additional code-gen workflow. But there is no associated test for
generated code for 'oneof' because the "exclusivity" verification has
not yet been code-generated.