VB-325: Add Kubernetes deployment example (#111)
This patch adds an example of deploying Concord (+Ethream RPC handler)
to a Kubernetes cluster. There is a disclaimer in the README.md
informing readers that this is an unsupported configuration,
product-wise.

The patch also modifies Concord's Dockerfile slightly to install
dnsutils (nslookup) in the runtime image.

In the example, each replica will expose its standalone ethrpc
endpoint to the rest of the K8s cluster.