Hermes: Make db-server probing more reliable
When hermes checks for the db-server container to be up and running it checks
whether a connection is possible. On my dev machine (Mac OSX) the first
connection attempt fails because the container isn't ready yet but the following
would be successful. However, once you have a failed connection the socket
object is pretty much unusable and it is recommended to close that socket and
create a new one. This change does exactly that plus minor refactoring.