Add protoc grpc compiler plugin
This patch adds GRPC codegen plugin for protobuf compiler for generating
code in Kotlin that is compatible with data classes generated with
protoc-plugin-kotlinx-serialization. The generated RPC stubs leverage
GRPC's Java runtime, as well as a newly created GRPC runtime support
library that is also included in this patch. The custom GRPC runtime
support makes sure that marshalling-related dependency by GRPC runtime
is leveraging kotlinx.serialization for marshalling/unmarshalling
functions.

This patch also adds a homogenous test to ensure that the generated RPC
server and client stubs work correctly (with itself).

Lastly, this patch attempts to cleanup / clarify the compile-time v.s.
runtime dependencies for the various components under the protobuf
project:

1. For protoc plugins, the packaging target is an executable JAR, and
should have only protobuf-java, kotlin-poet, and Kotlin platform
dependency as runtime dependency. The protobuf-java dependency is so
that the plugin has access to protobuf descriptor definitions available
at runtime. The kotlin-poet dependency is for the executable JAR to
output structured, generated Kotlin code. The Kotlin platform dependency
itself (in addition to implied JVM dependency) is for the executable
itself (to run on JVM).

2. For runtime-library that generated code depends on, the underlying
GRPC dependencies (grpc-core and grpc-stub) is declared as runtime
dependency since the runtime-library depends on them to function. Of
course JVM and Kotlin platform dependencies are also declared for
runtime dependency.

3. For the test library, every dependency is runtime (obviously for test
execution) but restricted to test-scope only.

For any projects that leverage the GRPC plugin, declaring (normal)
runtime dependency on the grpc-runtime-kotlinx-serialization will pull
in all the necessary transitive dependencies. (Although it is usually
good form to declare any dependency if the project has directly imported
code from a specific library.)

For any projects that only leverage the protobuf plugin (without GRPC),
it is necessary to explicitly declare the Kotlin platform dependency as
well as the kotlinx.serialization library dependency as runtime
dependencies.