VB-393: [UI] [security] Once logged into an instance, allows login to other instances with authentication
The issue is that we are using the same signing key in all instances of
Helen.  Any JWT will work on every instance of helen.

This changes generates a new random key every time helen reboots.  This
will prevent the behavior seen in the bug.

Testing: added some unit tests, and also tested by hand.  Prior to this fix, the UI continued to work between helen reboots.  Now when helen is rebooted, you are taken to the login screen.