Add preparatory steps for modular jar
This patch renames the runtime jar into a more "unique" package name in
order to prepare for exporting it as a module.

Java's JPMS requires modules to not overlap in the packages they export.
The previous package name for GRPC support is a bit general.

This patch also updates the Java version to 11 (to allow
module-info.java) to be created and bundled. module-info.java is not yet
added because the official io.grpc artifacts do not have module
descriptors yet, and has a problem with overlapping exported packages
itself.