[VB-991] Add ByteString for ByteArray serialization
This patch adds a ByteString implementation to act as the serialization
target for protobuf's 'bytes' scalar type.

ByteString represents an immutable sequence of bytes, and is implemented
in-memory as a wrapper around native ByteArray, with an offset and
length field denoting a sliced range of the buffer that the ByteString
should represent as its content (sequence of bytes).

Slice operation shares the underlying memory buffer with the origin
ByteString instance, and yields a new ByteString instance. Composition
of ByteString's is not yet supported.