Retry docker commands and consolidate docker logout commands
A server may be unavailable.  Retry failed docker commands to reduce
intermittently failing runs.

This change also consolidates the docker logout commands to always
execute when a run ends.

Retrying the curl commands is coming next. It will be a separate
submission because Curl can exit with 0 even in situations where
we would want to retry.