[VB-377] Add persistence and reactive components
This patch introduces a slim implementation of the Reactive Streams API
and uses them to define a shim asynchronous, non-blocking KV persistence
abstraction layer.

The KV persistence abstraction is implemented in 2 layers: typed and
untyped. The typed layer presents the traditional "mapped" value to the
upper layer, and takes care of marshalling/unmarshalling. The untyped
layer works with a "Value" functional interface type that knows how to
derive any type to a byte array.

The untyped KV abstraction is further implemented in 2 separate
components. The async interface is facilitated by
AbstractUntypedKeyValueStore, and takes care of the reactive streams
semantics. This is akin to a "client" component. The client component
delegates the actual operation asynchronously to an in-process server
component, by utilizing Channel constructs. Channel is a queuing
abstraction - depending on the type of channel, the Channel library uses
either full non-blocking ring buffer, wait-free non-blocking queue, or
unbuffered blocking (rendezvou) queue (of size 1). The "server"
component that is implemented in this patch utilizes a non-persisted
in-memory map. Pending the outcome of DB/datastore choice as well as the
persistence framework used, a preferably "server adapter" can be created
to make actual remote calls to the storage backend.

Finally the patcgh includes a set of unit-tests written to test the
various components introduced in the patch. There is a pure Java
unit-test to ensure the interoperability of the introduced code when on
the JVM and called from Java.