Format Concord: Wrapper script
We are interested in C/C++ related files only and hence we need to filter all
files under blockchain/concord. This change introduces a script that runs
clang-format with the proper arguments for you and is written with CI/CD in
mind. If you provide `--is-required` to the script then it will return 0 if no
code format change is needed and 1 otherwise.