Fix for retrying e2e tests on timeout
Since Jasmine doesn't report the filename on a timeout, we need to register and log the filename in a way the protractor-flake parser can understand. Based on this issue: https://github.com/NickTomlin/protractor-flake/issues/72 we override the global it function to register the filename and log it in a parser friendly way so that protractor flake will retry only the failing tests and not retry all tests when a timeout occurs.