concord-agent - First dummy implementation
Deployment service will have a component sitting in the concord-replica VM
which is responsible for:
1. Relaying status of deployed concord nodes back to Helen / fleet-management.
2. Relaying commands such as patching, stop/start (maintainance) from Helen to the deployed concord nodes.
3. Additional initial deployment handshake communication, etc.

This is the first basic skeleton which just builds using mvn.
There is NO functionality as of now.

Testing done:
mvn clean
mvn install

This is not integrated with docker-mvn yet (I haven't figured that one out yet)