[VB-377] Refactor reactive lib into separate module
This patch creates a core module under persephone (deployment service)
to house the reactive library code, which is generic and can be used by
all other sub-project components.

The intention of the core module is to house a small, light-weight,
internal-dependency-free set of core data types and library code for
other sub-project components.