[VB-375] Add Concord Model service
This patch adds the initial implementation for Concord Model service.
Specifically, the patch implements:

1. MetadataServer housing that also hosts the GRPC server (with Netty
as the socket server). For this patch, MetadataServer is wired up using
Google's Dagger compile-time dependency injection library. Although this
could be subject to change since it's not clear whether a simple
microservice such as this needs a dependency injection framework at all.
(i.e. inversion of control can be implemented via manual injection given
the current simplicity of the service components).

2. ConcordModelService that is minimally aware of GRPC's existence (in
the form of the RPC's service interfaces) and minimally aware of a
persistence storage backend (in the form of persistence KV interfaces).

The ConcordModelService instance is in control of its own lifecycle
command signals regardless of the actual server housing entity. For this
patch, the server housing entity is the MetadataServer.

Lastly, the patch implements a few unit-test cases that both directly
test against the ConcordModelService without GRPC, and with GRPC in the
form of a in-process GRPC server.