Modify protoc compiler plugin to emit Java-friendly accessor
Existing implementation of protobuf compiler plugin that emits Kotlin
data classes are meant to be 100% interop usable in Java code. But the
serializer accessor is only available to callers that can interpret
Kotlin's metadata in JVM bytecode (kotlinx.serialization emits the
bytecode in this way).

To make the serializer accessible to Java callers, an additional
@JvmStatic getter is code-generated in the companion object. So any Java
caller will be able to access the method like a static method for the
class, and be able to retrieve the serializer for any give such data
class.