Added support for YAML configuration files to configuration library.
This commit adds support for serializing and deserializing configuration parameters between ConcordConfiguration objects and YAML files. Two classes are added, YAMLConfigurationInput, which handles reading from a YAML file to a ConcordConfiguration, and YAMLConfigurationOutput, which handles writing a ConcordConfiguration to a YAML file. As with the other components of the configuration library, these classes are declared and documented in concord/configuration_manager.hpp, implemented in concord/configuration_manager.cpp, and unit tested in concord/test/config_test.cpp. This commit also introduces a dependency on yaml-cpp.