TxReceipt: Add more fields to log object
The motivation for this change is the web3 API and ethereum's JSON RPC
standard. While playing around with contracts, logs, and events the following
call failed because we return an incomplete log object:

    web3.eth.contract().events.Event().processReceipt(txr)

This API is a convenient wrapper in order to evalute events that happened on
the blockchain and it was looking for the 'logIndex' entry.