VB-602: implement ORIGIN command +test support
This change correctly sets "tx_origin" in the evm environment context,
so that the ORIGIN opcode produces the correct value.

This also required a test fix. The ethereum tests (sometimes?) specify
a "from" address for the transactions being executed. It's important
to use that value, because it shows up as data that is checked in the
result (this will be needed for VB-23 as well). This change gets
Hermes to use it.