getLogs: Concord
For now, we only support a specific block hash due to this change contains the
scaffoloding for the eth_getLogs request as well. End-to-end tests are in the
making and a getLogs call w/o parameters is planned as well in order to get all
logs from the latest block.