Make balance a 256bit type
It's been quite a journey but here is the result. The crux to understand
balance within concord is to internalize that its value is always going to be
in big endian format - no matter whether you provide/receive balance through
external API, look inside the EVM, or inside RocksDB. In this change, the only
time we convert to little endian is when we need to do math. The reason is that
we don't want to implement our own bit-twiddling but rather leverage the boost
library and its big number types.