VB-23: use DELEGATECALL to enable CALLER tests (#217)
* VB-23: use DELEGATECALL to enable CALLER tests

Long ago, we thought that CALLER wasn't implemented in our
system. After the correct account address was passed in (VB-602), it
became apparent that CALLER was implemented, but because of Hermes'
contract invocation indirection, the wrong CALLER was available to the
test. To get the correct CALLER to the test, Hermes needs to use the
DELEGATECALL opcode, instead of the CALL opcode. This is change number
1 in this PR (most of the changes to test_suite.py and
core_vm_tests.py).

Switching to DELEGATECALL uncovered the fact that we have implemented
DELEGATECALL incorrectly. When a contract is called as a delegate, its
code is supposed to run in the callee's environment - that is, it
reads and writes in the callee's storage. We were having it read and
write to its own storage. Fixing this is change number 2 in this PR
(all the changes in concord/src).

Finally, because some tests write test values to storage location
0x00, we needed a different location for Hermes to store the return
value of the code under test (since that 0x00 write is now in the
indirection contract's storage). I selected a location far outside the
test specs: 0x6865726d6573. That's "hermes" in ASCII, for those
wondering. Fixing this is change number 3 (the remaining changes in
test_suite.py and core_vm_tests.py).

Since CALLER now returns the correct value, the tests removed from the skip list by this
change pass. vmTests/arith.json and vmTests/boolean.json are reclassified as "23",
since they are no longer blocked on CALLER, but still pass when the
test definition says they should fail. This is change 4.

(All changes are part of one commit, just logically there are three
components.)

In the future, we should consider something @robem suggested: using
logs/events for return values instead of storage space. We might even
be able to remove the intermediate contract if we did this.