Rename build name of fleet-management to Persephone
There is sufficient amount of confusion over the name of Helen's Fleet
Management component. Internally, since the code base does not live
under helen/, another name was adopted to give the component a name.
"Persephone" has been the name that has been used since initial code
check-in.

This patch renames the build scripts and CI scripts so that the naming
of the artifact is consistent with the project component name.