getLogs: Filter by contract address
This change, combined with the previous getLog changes allows the user to
filter for events/logs from a specific contract within any block range.

- - - - >8 - - - - >8 - - - - >8 - - - - >8 - - - - >8 - - - - >8 - - - - >8

During testing I the web3 framework refused to work with non-checksummed
addresses which was built-in to protect the user. We should comply with that
approach and I filed VB-646 as a result.