Implement Persephone provision service
This patch implements the provision service API. The initial
implementation does not implement a persistent DeploymentLog. In fact,
the DeploymentLog source code is not yet utilized in the service code at
all. More changes are required to factor persistence and
async-write-ahead into DeploymentLog implementation.

As a result, there is a fair amount of CompletableFuture usage where
composed Publisher should be used instead. This will become clearer in
a follow up that properly implements DeploymentLog.

Jira: VB-639