getLogs: Filter by topic
This change adds the option to filter logs by topic. Depending on whether the
event arguments are indexed or not they appear in the data section or in the
topic section of the log entry. However, the first element in the topic list is
always the event signature. In our implementation we only produce one
transaction per block hence this change enhances the SimpleEvent.sol contract to
emit two events when the "twoEvents" function is called in order to test
multiple logs.