Enhance Persephone persistence generics typing
This patch refactors and cleans up previous generic typing declarations
to consistently constrain the T (Version) generic type to always be a
subtype of Version<T> (i.e. T extends Version<T>). Previously
KeyValueStore.Versioned as well as the command types in
UntypedKeyValueStore did not always constrain T to be a subtype of
Version, even though in usage T is always assumed to be a subtype of
Version.

This patch also fixes an imprecision typing issue with respect to
Version<T>.next(). Previously next() returns Version<T>. While this is
not incorrect, this is less precise than T, which means there are
use-site scenarios where an instance of T is guaranteed to yield another
instance of T through next(), but could not guarantee it due to the
relaxed type specifier.