Allow submodules to use git commands in dockerfile
Concord-bft has a requirement to pull in optional submodules via git
submodule update. The first optional submodule, for gtest was added and
merged. Unfortunately, the docker build doesn't work with the latest
concord-bft master because cmake gets run outside of a git directory.
The concord-bft submodule in concord/submodules/concord-bft has a .git
file with a relative path pointing to blockchain/.git. When we copy
concord out into the docker image, we also need to copy out the .git so
that it's in the same relative position and submodule update continues
to work from within cmake.

There are some alternatives to this, including running git sumbodule
update from the root (as is required currently) and then only pulling
optional submodules in concord-bft if the submodule directory is not
empty. This way the cmake will just ignore any already populated
submodules and not attempt to even run a git command. Unfortunately,
this strategy has complications, and will lead to broken builds during
development. As an example, let's say that the current sha of the
submodule for concord-bft is @abcdefg. When git submodule update --init
--recursive is run, it will update all existing submodules in
concord-bft for @abcdefg. However, when a developer working on
concord-bft checks out a new branch with new submodules, they won't
exist already. At this point, attempting a docker build will fail with
the error message `fatal: not a git repository:
/concord/submodules/concord-bft/../../../.git/modules/submodules/concord-bft`
or something similar. The developer will then have to understand this
message and know that they need to initialize submodules manually in
concord-bft, even though the build already does this for them normally,
without them even having to know about submodules.

In the end the strategy of manually initializing nested submodules, and
knowing about them, will lead to headaches and time spent debugging and
worrying about git, rather than allowing the developer to get on with
their job.  The one line change here can handle any submodule updates or
other git related commands in *any* submodule, including ones we don't
own.

Also, the .dockerignore files were modified to allow copying .git
directories and files.