Concord: Defaults accounts and contract defined error messages
While testing our supply chain dApp, we found some shortcomings in Concord.
This change expands the list of hardcoded default accounts from one to three
according to the genesis file. Given that we don't have an account story yet
this is an acceptable short term solution. The other change populates contract
defined error messages. In Solidity, a contract might use `revert` or `require`
- both take an error message as an argument. If the condition within `require`
is false or we hit `revert` then the EVM will revert the transaction and issue
EVM_REVERT as an error code. At that time we want to populate the contract
defined message if there is one.