Fix persephone logging artifact declaration
The logging facade that is used currently is slf4j. And for
metadata-server, the logging backend pulled in was supposed to be log4j2
to match what other Athena projects are using.

Due to the very confusing versioning scheme for log4j2 and the artifact
group ID, the wrong artifact was pulled in. As a result there was no
logging even though there was a "backend".

This patch specifies the correct artifact corresponding to log4j2. It
also "imports" the BOM for log4j2 in POM dependency management section
to avoid artifact version incompatibilities when using various
components of log4j2 together.