Add additional reactive stream helper functions
This patch adds helper extension functions to convert ReactiveStream's
Publisher to Java's CompletableFuture. The effect of converting a
Publisher turns the Publisher into a hot source (due to implicit
subscription to the publisher instance). In return, a CompletableFuture
is returned that completes when the Publisher finalizes the stream.

Note: Calling the converter on an infinite or potentially infinite
reactive stream is a very bad idea. Since the Future attempts to collect
all signals into a Collection, the end result would be
OutOfMemoryException. This is a similar mistake as attempting to
Stream.collect() on a Java 8 infinite Stream.