WIP - VB-640-part3
This is a placeholder file which will be used to package provisioning
service and config-utility into one.

Currently I have commented out most of the lines as it has dependency
on other code being checked-in.

This review that's why is WIP and needs to be re-worked once everything
is available but I don't want to loose what i have.

So generating review and if i can get quick review that it doesn't break
any existing thing, i would like to check-in.