Helen: eth_getTransactionByHash
This change copy & pastes (with import adjustments) the ethrpc implementation.
A couple of weeks ago I fixed the Ethrpc implementation but didn't update Helen
due to thought it wasn't used :( As clarified in MR106 we need both until we
have an authentication story for Ethrpc. Therefore, hermes' ExtendedRPCTests
should pass on on the Helen endpoint as well.

- - -

Testing Done:
* `ExtendedRPCTests`
* `ExtendedRPCTests --ethrpcApiUrl https://localhost/blockchains/local/api/concord/eth`

- - -

Follow-up: Adjust `gitlabBuildSteps.groovy` to run the test suite against Helen.