[VB-639] Reorganize provision service workflow
This patch reorganizes provision workflow such that public network
addresses are allocated upfront in advance of actual VM deployment. This
allows network addresses to be known to centralized config generation
utility to calculate every replica's config, before deploying the
replica VMs.

The reason for this is that without the persephone-agent in place, there
is no push-pull workflow to allow configuration files to be dropped in a
rendezvous location. Configuration for a node must be pushed upfront
during compute (VM) deployment as part of the deployment parameter set.

Lastly, this patch addresses a coroutine job cancellation bug that
causes child coroutines to be linked to the parent coroutines such that
cancellation propagates towards the parent and cancels the parent. This
is problematic for parents that are effectively singletons in the
system. In these scenarios, either child coroutines should be launched
in their own scope, or the coroutine context associated with the parent
should have a `Job` that is of supervisor type that is immune to child
coroutine failures. This patch takes the latter approach to properly
address the child coroutine failures to not cancel the parent as well.