VB-620: Integrate with Deployment service
Part one:  Provide the Helen APIs for deploying a blockchain.  This is
not integrated with the deployment service yet, but will let the UI team
start working on the UI

Updating the blockchain get format to correspond to feedback:
blockchain now returns a list of node entries.

Added a resource_link to the task information.

Updated tests.