Start with no blocks and add ability for tests to restart the product.
Stop adding blocks to the system as part of the startup routine.
If we monitor for Helen being able to return the nodes, and the nodes
all report they are live, then we have started.  There is no need
to do a test transaction.

Next, if a test case restarts the product, a new set of logs will
be created instead of overwriting the existing ones.

Finally, removing some dead code.