VB-743 Time Contract
This is the first version of the "time contract" required by the time
service for synchronizing time on the blockchain. Concord does not yet
use this contract. This change only gets the code in place to be used
in a future change.

This version of the time contract picks the median reading amongst the
latest values it has received. It does not yet enforce an approved
list of sources in any way, so it is possible to make time roll
backward by publishing enough "new" (previously unrecorded) sources.

The time contract expects to store its state in a KVB block under the
key 0x20 (i.e. 0x0220<block id>). Its state is encoded as the protobuf
message com::vmware::concord::kvb::Time.

Changes in build structure were made to allow the time contract test
code to use some concord modules without having to include their
sources explicitly. These changes are primarily to build
subdirectories as libraries, instead of including their sources in
libraries built by parent directories.