VB-744 include time with send-transaction requests
This change puts the previously-created time contract to use. If the
config `FEATURE_time_service` is enabled, replicas will use the time
contract for block time, instead of the timestamp in the EthRequest.

With the feature enabled, any concord node that has a `time_source_id`
will attach a TimeUpdate message to any ConcordRequest that contains
an ethereum transaction. Replicas will look for these TimeUpdates, and
apply them to their copy of the time contract at execution time.

The config `time_source_id` is optional, so that the number of sources
used can be tuned separately from the number of replicas. It is public
so that at some point in the future, the time contract can reject
updates from sources that haven't been registered previously.

None of the example config file has been altered. The nodes default
`FEATURE_time_service` to `false`, and therefore never look for a
`time_source_id`. The concord nodes behave as they always have before
this change.

`FEATURE_time_service` has been explicitly set to `false` in the
 config generator input files. This is not necessary, as the default
 value is `false` if not given. It's easier to modify configs for
 testing if the option is present, though.

`time_source_id` has been explicitly set to concord[1..4] in the
config generator input files. This is ignored if
`FEATURE_time_service` is disabled, but it's easier to test if the
option is present.

To test: edit one of the config generator input files to change
`FEATURE_time_service` to `true`, then generate new concord
configuration files using that input.