[VB-639] Add genesis and container registry to deployment workflow
This patch adds genesis.json model definition as a configurable
parameter to DeploymentSpecification.

The model itself is defined through protobuf. But Persephone API's
underlying serialization mechanism allows the generated protobuf
bindings to serialize/transcode to JSON. In this patch, the protobuf
definition is carefully defined to precisely match the genesis.json that
Concord currently takes.

This patch also makes container registry an injectable information to
the deployed VM. The source of this information is tied to the
OrchestrationSite that houses the VM. In other words, when
ProvisionService is started with the bootstrap config.json information,
the config would contain all the orchestration sites to be used by the
ProvisionService. each orchestration site info block will contain the
container registry that should be employed by that orchestration site -
typically this should be the container registry that is closest to the
region of the orchestration site. But in the most general case, it can
be a global load-balanced registry repository instead.