VB-379 - Install 'Agent' on the concord-node.
Today when deployment service deploys the concord-node (VM), it needs to be told which all docker images needs to
be downloaded explicitly. This will change once ModelService is implemented.

But for now, clients needs to indicate which Docker images are required to be installed on the host.

There are many things which needs to change but for now, we are using DockerHub for everything.

Testing done:
Deployed the concord-node using Deployment service and as you can see the images are copied into concord=node