Add checkstyle to persephone build
This patch is the first of many (4 currently) to apply the CheckStyle
rules as defined in Helen project to all persephone's Java code base.

This patch also modifies the checkstyle rule in 2 places:
1. Relaxes the JavaDoc line continuation indent comment spacing to 0 as
it actually improves readability of the source code without actually
inducing any changes to the generated javadoc comment.

2. Relaxes copyright preamble and postamble slightly to allow trailing
"*****" line comments to sandwich the actual copyright statement. The
style is used elsewhere in VMware with great frequency, and is included
in various style guides as referenced by the official coding standards
wiki at https://wiki.eng.vmware.com/JavaCodingStandard .