run git checkouts with their clones
Putting the checkout in the same RUN as the clone means that if we
ever change the hash we're checking out, docker will be forced to also
re-clone the repo. This ensures that newer commits get pulled in,
instead of allowing an old cached clone to cause the build to fail.