Incorporate persephone and protobuf into master build
This patch simply adds persephone and protobuf into the master build
manifest (top level maven pom). Note that the build order listed in the
POM is not lexicographical because the order will actually populate the
build-host's local repository with the necessary inter-dependencies.

This implicit dependency is in fact high *undesirable*. But until
buildall.sh stops relying on the top-level maven build POM to perform
the build, this patch is necessary.