Changes Ethereum Proof of Work Nonces to 8 Bytes.
This commit changes the Ethereum proof of work nonces that Concord returns (generally to EthRPC or Helen) from 32-byte byte arrays to 8-byte byte arrays to be more consistent with Ethereum, which uses 8-byte nonces for proof of work. Note this commit does not change the fact that proof-of-work nonces are functionally unused by Concord and just set to 0.