[Temporary] Additional fixes to enable Concord to use private IP address
This patch continues to make additional concessional changes to allow
Concord to use private network addresses instead of public IP address in
order to workaround current VMC NSX networking limitations.

In addition, this patch fixes an issue in cloud init script where
Concord and ethrpc on the same node cannot talk to each other due to
ethrpc expecting Concord to be "localhost". Concord will only be
"localhost" if host networking is used instead of Docker bridge network
(which is the default if --network is unspecified during docker run).