Concord: Build DAML PoC protobuf with v8
With MR !173 we updated Concord's dependecies to use Golang (for HLF) and gRPC
(HLF and DAML). This change makes use of the gRPC dependency and is the first
MR for the DAML PoC. The protobufs are compiled but not used yet which will
come in a later change. In order to compile, I leveraged @jiahaoc's work in
MR !153 and separated it out for this change.