Concord: Add DAML service and commands handler
In this change I introduce the DAML code from the prototype - the parts that
are used to build a KV service on top of a consensus algorithm. On one hand we
have `grpc_services` which defines all gRPC endpoints for the API server and on
the other hand there is the commands handler which gets invoked when consensus
is reached. Most notably, all incoming requests are packed within a
ConcordRequest/*Response and further into a DamlRequest/*Response type. The use
of the API services and the new commands handler will be implemented with the
next change.

Note: The code is mostly copy & paste with adaptions.

FYI, in here you can see how we integrate protobuf3 messages within version2.
Converting concord.proto to v3 wasn't as straight forward hence the workaround.